feat: add implement swapDYfiToYfi

+++ src/interfaces/curve/ICurveTwoAssetPool.sol
@@ -12,8 +12,8 @@ interface ICurveTwoAssetPool {
         returns (uint256);
 
     function add_liquidity(uint256[2] calldata amounts, uint256 minMintAmount) external returns (uint256);
-    function get_dy(int128 i, int128 j, uint256 dx) external view returns (uint256 dy);
+    function get_dy(uint256 i, uint256 j, uint256 dx) external view returns (uint256 dy);
     // solhint-disable-next-line func-param-name-mixedcase,var-name-mixedcase
-    function exchange(int128 i, int128 j, uint256 dx, uint256 min_dy) external;
+    function exchange(uint256 i, uint256 j, uint256 dx, uint256 min_dy) external;
     function coins(uint256 arg0) external view returns (address);
 }
